requirements.txt
Django==1.10
django-angular==0.8.4
djangorestframework==3.4.6
psycopg2==2.6.2
